arm64: DT: SDM660: Migrate audio DT to k4.14 and fix DTC warnings
The audio on kernel 4.14 has got a new structure in the device-tree:
now we probe the Q6 Core driver and that one will be responsible
for probing the rest of the audio drivers. This is truly sane, since
everything has a dependency on that one and major audio related
communication happens through the Q6.

While at it, solve most of the DTC compiler warnings, so that we
get a nice clean output.